+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Confluent's Python client for Apache Kafka
 
 
-## v1.10.0
+## v1.9.3
 
-- Add metadata to TopicPartition type and commit() (#1410).
+ - Added metadata to `TopicPartition` type and `commit()` (#1410).
+ - Added `consumer.memberid()` for getting member id assigned to 
+   the consumer in a consumer group (#1154).
 
 
 ## v1.9.2